EISW-25644 KW fixes for the ww43 release (#110)
* KW Fixes

Fixes Applied (10 issues):
	5443: [lp_scheduler] Changed assert to throw runtime error
	8143, 8144, 8145, 8147, 8149, 8151, 8154, 8155: [heuristic_strategy] Moved "init" function to class constructor
	8472: [pass_utils] Added check for pointer validity to if statement

* Numerous KW fixes

- uninitialized variables in constructor
- void functions returning values
- incorrect param type (size_t) to printf
- Correctly close file resources

* Klockwork fixes for file closing and uninit vars

* Fixes addressing KW issues

* Klocwork fixes up to #9891

* Klocwork fixes up to issue #9008

* KW fixes batch

* Additional KW fixes

* KW fixes, batch 4

* KW fixes batch 5

* KW fixes

* Addressing comments from MR# 3471

* Addressing comments and failing unit tests

* Backing out a KW fix - not required

* Reverting KW updates - these will be marked not a problem in KW

* Reverting KW fix - will mark as not a problem

* Reverting KW fix as not required

+++ src/mcmCompiler/src/pass/lp_scheduler/pipeline_chains_transform.hpp
@@ -81,7 +81,7 @@ class Pipeline_Chains {
           operation_t dpu_op = *dpu_op_itr;
           fprintf(fptr, "%s :  reads=%zu output=%zu",
               (dpu_op->getName()).c_str(),
-              total_read_size_map_[dpu_op], cmx_size(dpu_op));
+              (size_t)total_read_size_map_[dpu_op], cmx_size(dpu_op));
           fprintf(fptr, "\n");
 
           {
